# services/api.tapped.ai/src/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tapped API Docs",
    "summary": "the leading API for live music data including performers, venues, and events. High quality live music data and aggregates",
    "description": "Errors share one shape: `{ \"error\", \"error_id\", \"error_details\"? }`. Quote `error_id` when reporting a problem; it is logged server-side.",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "https://api.tapped.ai"
    }
  ],
  "paths": {
...
    "/app/v1/data/{table}": {
      "get": {
        "tags": [
          "app"
        ],
        "summary": "List documents",
        "parameters": [
          {
            "in": "query",
            "name": "after",
            "schema": {
              "type": [
                "string",
                "null"
              ]
            },
            "style": "form"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64"
            },
            "style": "form"
          },
          {
            "in": "query",
            "name": "mode",
            "schema": {
              "type": [
                "string",
                "null"
              ]
            },
            "style": "form"
          },
          {
            "in": "query",
            "name": "order",
            "schema": {
              "type": [
                "string",
                "null"
              ]
            },
            "style": "form"
          },
          {
            "in": "query",
            "name": "referenceEventId",
            "schema": {
              "type": [
                "string",
                "null"
              ]
            },
            "style": "form"
          },
          {
            "in": "query",
            "name": "requesteeId",
            "schema": {
              "type": [
                "string",
                "null"
              ]
            },
            "style": "form"
          },
          {
            "in": "query",
            "name": "requesterId",
            "schema": {
              "type": [
                "string",
                "null"
              ]
            },
            "style": "form"
          },
          {
            "in": "query",
            "name": "reviewType",
            "schema": {
              "type": [
                "string",
                "null"
              ]
            },
            "style": "form"
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": [
                "string",
                "null"
              ]
            },
            "style": "form"
          },
          {
            "in": "query",
            "name": "userId",
            "schema": {
              "type": [
                "string",
                "null"
              ]
            },
            "style": "form"
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppError"
                },
                "example": {
                  "error": "some error happened",
                  "error_id": "00000000-0000-0000-0000-000000000000"
                }
              }
            }
          },
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "FirebaseAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "app"
        ],
        "summary": "Create an owned document",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          },
          "required": true
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppError"
                },
                "example": {
                  "error": "some error happened",
                  "error_id": "00000000-0000-0000-0000-000000000000"
                }
              }
            }
          },
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "FirebaseAuth": []
          }
        ]
      }
    },
...
    "/app/v1/stream-token": {
      "post": {
        "tags": [
          "app"
        ],
        "summary": "Mint a Stream chat token for the signed-in user",
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppError"
                },
                "example": {
                  "error": "some error happened",
                  "error_id": "00000000-0000-0000-0000-000000000000"
                }
              }
            }
          },
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StreamUserTokenResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "FirebaseAuth": []
          }
        ]
      }
    },
    "/app/v1/places/photo": {
      "get": {
        "tags": [
          "app"
        ],
        "summary": "Resolve a Google place photo to a short-lived URL",
        "parameters": [
          {
            "in": "query",
            "name": "maxHeightPx",
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "uint32",
              "minimum": 0.0
            },
            "style": "form"
          },
          {
            "in": "query",
            "name": "name",
            "required": true,
            "schema": {
              "type": "string"
            },
            "style": "form"
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppError"
                },
                "example": {
                  "error": "some error happened",
                  "error_id": "00000000-0000-0000-0000-000000000000"
                }
              }
            }
          },
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhotoResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "FirebaseAuth": []
          }
        ]
      }
    },
    "/app/v1/places/reverse-geocode": {
      "get": {
        "tags": [
          "app"
        ],
        "summary": "Locality place for a coordinate",
        "parameters": [
          {
            "in": "query",
            "name": "lat",
            "required": true,
            "schema": {
              "type": "number",
              "format": "double"
            },
            "style": "form"
          },
          {
            "in": "query",
            "name": "lng",
            "required": true,
            "schema": {
              "type": "number",
              "format": "double"
            },
            "style": "form"
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppError"
                },
                "example": {
                  "error": "some error happened",
                  "error_id": "00000000-0000-0000-0000-000000000000"
                }
              }
            }
          },
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReverseGeocodeResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "FirebaseAuth": []
          }
        ]
      }
    },
    "/app/v1/places/locality": {
      "get": {
        "tags": [
          "app"
        ],
        "summary": "Locality place details for a coordinate",
        "description": "`reverse-geocode` and `places/{placeId}` in one request; 404 when no locality contains the coordinate.",
        "parameters": [
          {
            "in": "query",
            "name": "lat",
            "required": true,
            "schema": {
              "type": "number",
              "format": "double"
            },
            "style": "form"
          },
          {
            "in": "query",
            "name": "lng",
            "required": true,
            "schema": {
              "type": "number",
              "format": "double"
            },
            "style": "form"
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppError"
                },
                "example": {
                  "error": "some error happened",
                  "error_id": "00000000-0000-0000-0000-000000000000"
                }
              }
            }
          },
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaceDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "FirebaseAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "tapped-api-key",
        "description": "your API Key"
      },
      "FirebaseAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "A Firebase Auth ID token for the tapped project."
      }
    },
    "schemas": {
...
      "AppError": {
        "description": "The error body every route returns: `{ \"error\", \"error_id\", \"error_details\"? }`.",
        "type": "object",
        "required": [
          "error",
          "error_id"
        ],
        "properties": {
          "error": {
            "description": "An error message.",
            "type": "string"
          },
          "error_details": {
            "description": "Optional Additional error details."
          },
          "error_id": {
            "description": "A unique error ID, also written to the server logs.",
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "PhotoResponse": {
        "type": "object",
        "properties": {
          "photoUri": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "ReverseGeocodeResponse": {
        "type": "object",
        "properties": {
          "placeId": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "StreamUserTokenResponse": {
        "type": "object",
        "required": [
          "token"
        ],
        "properties": {
          "token": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}